// src/common/enums.d.ts
export declare enum DIRECTION {
	ROW = 'row',
	COLUMN = 'column',
}
export declare enum MARGIN_SIDES {
	TOP = 'top',
	RIGHT = 'right',
	BOTTOM = 'bottom',
	LEFT = 'left',
}
export declare enum ALIGN_ITEMS {
	NORMAL = 'normal',
	FLEX_START = 'flex-start',
	FLEX_END = 'flex-end',
	CENTER = 'center',
	BASELINE = 'baseline',
	STRETCH = 'stretch',
	START = 'start',
	END = 'end',
	SELF_START = 'self-start',
	SELF_END = 'self-end',
	FIRST_BASELINE = 'first baseline',
	LAST_BASELINE = 'last baseline',
	SAFE_CENTER = 'safe center',
	UNSAFE_CENTER = 'unsafe center',
	SAFE_RIGHT = 'safe right',
	UNSAFE_RIGHT = 'unsafe right',
	SAFE_END = 'safe end',
	UNSAFE_END = 'unsafe end',
	SAFE_SELF_END = 'safe self-end',
	UNSAFE_SELF_END = 'unsafe self-end',
	SAFE_FLEX_END = 'safe flex-end',
	UNSAFE_FLEX_END = 'unsafe flex-end',
}
export declare enum JUSTIFY_CONTENT {
	START = 'start',
	END = 'end',
	FLEX_START = 'flex-start',
	FLEX_END = 'flex-end',
	CENTER = 'center',
	LEFT = 'left',
	RIGHT = 'right',
	SPACE_BETWEEN = 'space-between',
	SPACE_AROUND = 'space-around',
	SPACE_EVENLY = 'space-evenly',
	STRETCH = 'stretch',
}
export declare enum SIZE_UNITS {
	AUTO = 'auto',
	PERCENT = '%',
	FN = 'fn',
	PX = 'px',
	FR = 'fr',
	REM = 'rem',
	VH = 'vh',
	VW = 'vw',
	PC = 'pc',
	CM = 'cm',
	MM = 'mm',
	IN = 'in',
	PT = 'pt',
	CH = 'ch',
	EM = 'em',
	V_MIN = 'vmin',
	V_MAX = 'vmax',
}
export declare enum COLOR_MODE {
	DARK = 'dark',
	LIGHT = 'light',
}